% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{bw_level}
\alias{bw_level}
\title{Compute the score for a specific bandwidthfunction}
\usage{
bw_level(bw, nvalsets, d, pts, xmax, xmin, ymax, ymin)
}
\description{
Compute the score for a specific bandwidthfunction
}

