% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{d2}
\alias{d2}
\title{Compute the distance between two points}
\usage{
d2(a, b)
}
\arguments{
\item{a,b}{2-dimeninsional vectors}
}
\description{
Compute the distance between two points
}

