% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{exp_value}
\alias{exp_value}
\title{Compute the exponentiated part of the bandwidth optimizer}
\usage{
exp_value(validationRegion, validationRegionArea, checkedRegion, checkedArea,
  checkset, bw)
}
\description{
Compute the exponentiated part of the bandwidth optimizer
}

