% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{final_comp}
\alias{final_comp}
\title{Compute the full bandwidth optimizer for a single validation set}
\usage{
final_comp(validSet, checkedSet, validationRegion, validationArea,
  checkedRegion, checkedArea, bw)
}
\description{
Compute the full bandwidth optimizer for a single validation set
}

