% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda_hat.R
\name{in_Area}
\alias{in_Area}
\title{Discover whether points are in a given region}
\usage{
in_Area(pts, region)
}
\arguments{
\item{pts}{A two-column matrix with x-values in column 1 and y-values in column 2.}

\item{region}{A set containing information on validation set(needs to be removed),
lattice points, 1-st tree neighbor, and distance from lattice to tree neighbor}
}
\description{
Discover whether points are in a given region
}

