% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{inCheckedRegion}
\alias{inCheckedRegion}
\title{Determine which points are in the checked region}
\usage{
inCheckedRegion(pts, checkset)
}
\description{
Determine which points are in the checked region
}

