% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{inValidationRegion}
\alias{inValidationRegion}
\title{Determine which points are in the validation region}
\usage{
inValidationRegion(pts, valset, checkset)
}
\description{
Determine which points are in the validation region
}

