% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KDE_Function.R
\name{KDE_1}
\alias{KDE_1}
\title{General One-Dimensional Gausssian Kernel Density Estimate}
\usage{
KDE_1(d, bw = my_ucv(d)$par, dmin = min(d) - bw * abs(pnorm(0.025,
  min(d)/bw)), dmax = max(d) + bw * abs(pnorm(0.975, max(d)/bw)),
  accuracy = 1000)
}
\arguments{
\item{d}{A vector of observations.}

\item{bw}{Bandwidth. Default is the unbiased cross-validation estimate.}

\item{dmin}{The lowest density you'd like to estimate. Default somewhat
arbitrary. Used two-standard deviations from the relevant value.}

\item{dmax}{The highest quantity you'd like to estimate. Default somewhat
arbitrary. Used two-standard deviations from the relevant value.}

\item{accuracy}{Numer of bins you'd like. Default somewhat arbitrary.}
}
\value{
A matrix with locations and heights.
}
\description{
Computes the one-dimensional guassian KDE for an entire region.
}
\examples{
x <- rnorm(100)
a <- KDE_1(x)
plot(a, type ='l')

y <- rgamma(30, 2, 4)
KDE_1(x, 0.25)
b <- KDE_1(y, 5)
c <- KDE_1(y)
plot(b, type = 'l')
lines(c, type = 'l', col = 'red')
}

