% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KDE_Function.R
\name{KDE_height}
\alias{KDE_height}
\title{Single-Point One-Dimensional Gausssian Kernel Density Estimate}
\usage{
KDE_height(p, d, bw)
}
\arguments{
\item{p}{A real number}

\item{d}{A vector of observations}

\item{bw}{Bandwidth}
}
\value{
A numeric density estimate
}
\description{
Computes the one-dimensional guassian KDE for a single point.
}

