% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda_hat.R
\name{lambda_A_checked}
\alias{lambda_A_checked}
\title{Calculates the denominator for lamda_hat}
\usage{
lambda_A_checked(pt, my_region, my_Area, bw)
}
\description{
Calculates the denominator for lamda_hat
}

