% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda_hat.R
\name{lambda_hat}
\alias{lambda_hat}
\title{Calculate non-parametric intensity of an IPP at a point}
\usage{
lambda_hat(pt_x, pt_y, my_set, bw, my_region, my_Area)
}
\arguments{
\item{pt_x}{An x value}

\item{pt_y}{A y value}

\item{my_set}{A set containing information on validation set(needs to be removed),
lattice points, 1-st tree neighbor, and distance from lattice to tree neighbor}

\item{bw}{The desired bw.}

\item{my_region}{The portion of the space already explored}

\item{my_Area}{The area of the explored space}
}
\value{
A non-parametric estimate of the intensity at the point (pt_x, pt_y)
}
\description{
Calculate non-parametric intensity of an IPP at a point
}

