% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BW_Cross_Validation.R
\name{my_bcv}
\alias{my_bcv}
\title{Biased Cross-Validation for One-Dimensional Gaussian Kernel Bandwidth
Selection.}
\usage{
my_bcv(d, lower = 0.0025, upper = my_oversmooth(d))
}
\arguments{
\item{d}{A vector.}

\item{lower}{A real number.}

\item{upper}{A real number.}
}
\value{
Computes the biased cross-validation bandwidth of a one-dimensional
  gaussian kernel density estimator.
}
\description{
Biased Cross-Validation for One-Dimensional Gaussian Kernel Bandwidth
Selection.
}

