% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BW_Cross_Validation.R
\name{my_oversmooth}
\alias{my_oversmooth}
\title{Oversmoothed bandwidth Estimate for One-Dimensional Gaussian KDE}
\usage{
my_oversmooth(d)
}
\arguments{
\item{d}{A vector}
}
\value{
Computes the oversmoothed bandwidth estimate for the one-dimensional
  Guassian kernel density estimate.
}
\description{
Oversmoothed bandwidth Estimate for One-Dimensional Gaussian KDE
}

