% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BW_Cross_Validation.R
\name{my_ucv}
\alias{my_ucv}
\title{Unbiased Cross-Validation for One-Dimensional Gaussian Kernel Bandwidth
Selection.}
\usage{
my_ucv(d, lower = 0.0025, upper = 2)
}
\arguments{
\item{d}{A vector.}

\item{lower}{A real number.}

\item{upper}{A real number.}
}
\value{
Computes the unbiased cross-validation bandwidth of a one-dimensional
  gaussian kernel density estimator.
}
\description{
Unbiased Cross-Validation for One-Dimensional Gaussian Kernel Bandwidth
Selection.
}

