% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{nearest_point}
\alias{nearest_point}
\title{Find the data point nearest to (x,y)}
\usage{
nearest_point(x, y, d)
}
\arguments{
\item{x,y}{Real numbers.}

\item{d}{A two column  matrix with each row a point}
}
\description{
Find the data point nearest to (x,y)
}

