% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda_hat.R
\name{num_lambda_hat}
\alias{num_lambda_hat}
\title{Calculates the numerator of the non-parametric intensity estimate}
\usage{
num_lambda_hat(pt, my_set, bw)
}
\description{
Calculates the numerator of the non-parametric intensity estimate
}

