% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{opt_bw}
\alias{opt_bw}
\title{Find the optimal bandwidth.}
\usage{
opt_bw(npts, nvalsets, d, xmax, xmin, ymax, ymin, min_guess, max_guess,
  step_value)
}
\description{
Currently to slow to work well
}

