% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{prob_O_Si}
\alias{prob_O_Si}
\title{Perform the whole validation set bandwidth probability computation from start to finish}
\usage{
prob_O_Si(val_num, d, pts, bw, xmax, xmin, ymax, ymin)
}
\description{
Perform the whole validation set bandwidth probability computation from start to finish
}

