% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{prod_part}
\alias{prod_part}
\title{Compute the product part of the bandwidth optimizer}
\usage{
prod_part(valid_set, checkedRegion, checkedArea, checkset, bw)
}
\description{
Compute the product part of the bandwidth optimizer
}

