% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvdnorm.R
\name{twod_dnorm}
\alias{twod_dnorm}
\title{Bivariate Guassian Density Computation Compute the density of a point under a
bivariate gaussian model}
\usage{
twod_dnorm(p, mu, sigma)
}
\arguments{
\item{p}{A (x,y) vector of real numbers, i.e. a point.}

\item{mu}{A vector indicating the mean.}

\item{sigma}{Covariance Matrix}
}
\value{
The density at point p of a bivariate gaussian with mean mu and
  covariance matrix sigma.
}
\description{
Bivariate Guassian Density Computation Compute the density of a point under a
bivariate gaussian model
}

