% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Paper_BW_Cross_Validation.R
\name{val_sets}
\alias{val_sets}
\title{Create a vector with validation set information}
\usage{
val_sets(sets_number, total)
}
\arguments{
\item{sets_number}{How many validation sets.}

\item{total}{Total number of items.}
}
\description{
Create a vector with validation set information
}

